MODFEE-146: Store new aged to lost settings for recalled items
I added the requested fields to the RAML file for this policy type
this seems to make the build process automatically generate the
changes to the model and the underlying database.  I also inserted
the new fields in the lost item policy API test file.  Since this
file doesn't seem to check that any specific field exists in any
created policy, I didn't put in anything that specifically checks
for these fields either. However, I did confirm that that the fields
were present in the resulting posted policy by dumping the entire
response to screen and looking at it.  I also confirmed that if the
API was not expecting those fields or able to store them, it would
respond with a 422 error, which is not the case.

refs: https://issues.folio.org/browse/MODFEE-146